LayoutTests/platform/ios/imported/w3c/web-platform-tests/selection/selection-direction-on-triple-click.tentative-expected.txt

hello, world

FAIL direction returns "none" when there is a triple click selection(directionless) assert_equals: expected Element node <div id="container">hello, world</div> but got null
